DAQ-2559 Allow step and points to be used for 'grid' scans
Currently the word 'raster' is used to denote a scan over a grid of
points defined by specifying the spacing between them along two axes.
The word 'grid' is used to denote a grid scan over a grid of points
defined by specifying the number of points along two axes within a
bounding box determined by the specified ROI. Both these paths are in
fact raster scans and grid scans in reality which is confusing.

A better nomenclature would be to specify them as step scans or scans
over a number of points (in the same way that you can for a line). This
could then be consistently applied to scans regardless of the number of
axes over which they operate with the only requirement then being that
for each axis the step size or no of points must be specified. The
exception to this is the line in any number of axes where the step and
number of points apply along the line itself.

This change implements this behaviour for the existing scan paths. This
will hopefully provide a consistent and painless way of moving away from
previous poorly chosen nomenclature. In addition new aliases are added
for line and axial scans to allow these to be expressed in a form that
is more meaningful for tomography.

Change-Id: If041143171cf3091eb63269ec152d23ed51df7dd