I20-471 : Set 'Exafs step type' to 'k' in Xas parameters GUI.
This change allows the 'Exafs step type' to be fixed 'k' in the 'Xas parameters GUI'
and the combo box for choosing it disable. This can help prevents a user
from accidentally changing the step type to 'E' (energy) and scan parameters with
10000+ energy points in it.
This behaviour can be switched on my setting the plugin customization preference
exafs.only.allow.kstep.type.Preference to 'true'.

Change-Id: Id49c1addfd1bea6867be8e9a7b780edb75d70e65