set duration for dummy point generator in MalcolmDevice
The dummy point generator is used to validate a malcolm model outside
of a scan, i.e. where we just want to validate the detector models.
Since we're not in a scan, there is no point generator, but malcolm
required a generator to be present in the EpicsMalcolmModel. To handle
this we use a dummy point generator created from a StaticModel.
It seems that malcolm now checks that the exposure time is set to a
positive value, so we set it to 0.1 (the actual value doesn't matter)

Change-Id: I14c3163caa221c081687ece615cbb2e046990936