[DAQ-2477] remove spiral generator structure from block PV structure,
A real malcolm device does not have a 'generator' field holding the
last point generator. I understand that it was added to test
serialization and deserialization of point generators. However, the
current tests don't actually test that. Future changes for this ticket
will add deserialization of point generators with appropriate tests.

Change-Id: I7bcf8f45ea058264c326dbf8adefa64d4af32086