copy the CompoundModel to ensure we're not cheating
The expectedConfigureMessage and expectedValidateMessage that the
epics connection Mockito mock was set up with using both when and validate
was incorrect at the time it was set up. However we were getting away
with this because the CompoundModel used was the same instance that is
configured in MalcolmDevice.createEpicsMalcolmModel. So when that method
set the duration and mutators, this also made the expected model
correct. This was kind of cheating. So we now copy the CompoundModel
and explicitly set the duration and mutators so we can check that
MalcolmDevice sets these as expected.

Change-Id: Ie2e4736fc0f80a72152b04e6d16fe31a909fc9c6