DAQ-2746 Add spring motor parser
This enables the automatic, mode aware creation of motors in Spring
configurations. This replaces the need for three beans to be created in
different directories and kept in sync when anything changes.

Where previously a motor would be defined similarly to

    <!-- in common config -->
    <bean id="stage_x" class="gda.device.scannable.ScannableMotor" >
        <property name="motor" ref="stage_x_motor" />
    </bean>

    <!-- in live config -->
    <bean id="stage_x_motor" class="gda.device.motor.EpicsMotor" >
        <property name="pvName" value="BL22I-MO-STABL-01:X" />
    </bean>

    <!-- in dummy config -->
    <bean id="stage_x_motor" class="gda.device.motor.DummyMotor" />

It can now be defined using a single declaration

    <gda:motor id="stage_x" pv="BL22I-MO-STABL-01:X" />

where the gda namespace is the existing namespace used to provide RMI
imports and exports.

Most of the common additional customisation that motors require can be
added as attributes to the motor element, eg:

    <gda:motor id="stage_x"
            userUnits="um"
            protectionLevel="3"
            pv="BL22I-MO-STABL-01:X" />

with the properties being set on the scannable motor or the motor
controller as appropriate.

If further customisation is required, the motors can continue to be defined
in the current style although if a feature is required often, this
mechanism could be extended to include it.

The classes used for live and dummy mode can be configured, either by
setting the system wide default via config_instance.properties or on a
case by case basis via dummy/liveImplementation attributes. If nothing
is set in either place, the most common current classes are used
(EpicsMotor and DummyMotor).

Change-Id: I0ad8917f84bceaad62984b99c6780f561a65c04b