DAQ-2723 Add start, stop and end of scan hooks to Nexus Processor related interfaces
For some use cases, e.g. SWMR file opening, need to close the file at the
end of a scan or if the scan is stopped.

For processors which extract ROI data from a gui - this should be done
be done at the start of a scan to ensure consistency.

Change-Id: I8fb028c9bdd7d5eb9b04a2c2d610529fa426b654