DAQ-2778 Refactor and extend DatasetStats
Add new statistics and group all these together in an enum. This provides
default output names for each statistic but this can be configured if
desired. The set of statistics which are actually used is also configurable
with sum and average being the defaults as before.

This also fixes DAQ-2780 (incorrect names in the NXDetectorData) and
DAQ-2779 (statistic data should be point dependent).

Change-Id: I8c2324e871bb95b1bd30d95f9f8d1f830534b1aa