DAQ-2637 Allow different widget types in LinkContributionItem
LinkContributionItem is used to add widgets to the status line.
Previously it did this with a CLabel, but there is a requirement
to add a button in the same way - and other types of widget may
be required in future.

CLabel and Button define a similar set of functions, but they are
not actually derived from a common base class.

This change adds an interface LinkContributionWidget containing the
functions that LinkContributionItem calls, and implementing classes
that wrap a CLabel & Button respectively.

Change-Id: Ie8b4bde9e3d0e5c2608c9ae62f5c785a99eaf807