B07-414 Buttons hidden in Alignment Controls view in SPECS perspective
When the alignment controls view is made smaller vertically, the Start, Set and Stop buttons disappear behind the Intensity label and text box. No scroll bars appear so it is not apparent that there is hidden content, and there is no way to access the content without resizing the pane.

Change-Id: I897e37ddcde59f931e055be8defb0fc80ff5fc13