[DAQ-2765] handle scanning queue loading
Currently if the mapping queue cannot be deserialized this will throw
an exception, causing the GDA servers to fail to start. This situation
can occur when switching between versions of GDA where there have
been changes to the points models, e.g. between GDA 9.15 and 9.16
many of the points models, such as GridModel, have been renamed.

The solution is to catch the exception, delete the existing database
file used by MVstore and try again. This will cause MVstore to try
again with a new empty database file.

Change-Id: Ia860c335720210150948ec118789c6abc322208d