DAQ-2793 Remove org.monte.media plugin from GDA
This used to be a dependency of org.dawnsci.conversion but this has
been removed.

The removal from org.dawnsci.conversion meant that DAWN no longer
depended on the dawn-third repository. This meant that it was not
tagged with a new branch/tag so the GDA 9.16 build failed as it still
had this dependency.

Change-Id: I4a47e7725aafc176a865bd63b64900cd1a78b142