B18-463 : Make collectionTime settable in MultipleExposureHarwareTriggeredStrategy
This is used for setting the detector frame time for each data point in
B18's qexafs scans with Medipix detector. Previously this class was used only
on I20, where the collection time has no effect since it is determined
by the length of the TTL trigger pulse from the Tfg.

Change-Id: Ic63f95a45d41f51213660861bae9a30d4a7005d1