K11-592 Create index file at end of experiment
When an experiment is closed, an index file is created in the experiment
directory which links to every acquisition that occurred within its
duration. Each acquisition appears as a separate entry with the same
name as the acquisition itself.

Eventually - not implemented in this change - the file will also contain
the relevant metadata for the experiment e.g. beam energy, sample name/
description, detector position, etc.

Creation of the index file is accomplished via a request/response
mechanism over ActiveMQ. This change therefore introduces a responder
for all jobs related to the experiment structure; IndexFileCreation-
Request is the first implementation of such a job. Any new type of job
must also have a handler mapped to in in the ExperimentStructureJob-
HandlerFactory.

Change-Id: I6cdb76de75612ef98ee9bce8a5cf70f794db6533