[DAQ-2765] iterate over submission and status queues after loading
in order to check that an exception isn't thrown. This was an issue
that Anthony was experiencing when switching between GDA 9.15 and 9.16.

Add a test for this by using override the iterator() method of the queue
class to throw a RuntimeException. Note that it was done this way
because it was proving very difficult if not impossible to achieve this
with mockito/powermock for unknown reasons.

Change-Id: Icb0289f8c2cc964a1a38b1bcbeb06e63530f9c78