Add initial version of gda-conf.py
This is a small python app which enables GDA to be started
using a set of spring profiles chosen from a checkbox gui.

The list of profiles (and their default status) is defined
in a config file which looks like:

 [options]
 beamline = i11

 [devices]
 hiden_rga = True
 pixium_detector = True
 ms3_stage = False

The path to this file is passed as an argument to the script.

Note that core-config/bin is not on the PATH for beamline
workstations so an entry must be made in the launcher
json for the beamline.

It has already been in use on I22 but is being placed
here so that it can be used easily on other beamlines.

This file has been imported from https://gitlab.diamond.ac.uk/snippets/8
which is an updated version of:
gda-diamond.git/configurations/i22-config/bin/gda-conf
added in change #15481 that should be eventually replaced with this.

Change-Id: I2d91bf15707c7f8f7b8317e8fc5139d76c605367