DAQ-2783 Allow user to specify units for "Scale" when defining spiral mapping path
The "scale" of a spiral scan actually controls two things:

(1) the spacing of the "rings" of the spiral, so a small scale gives
you a tightly-wound spiral, and a large scale the opposite

(2) the spacing of points along the line of the spiral

In practice, the value given is approximately the spacing in mm between
points as drawn in the Map view. This change gives the user the ability
to specify the value in different units. It also allows you to
specify what units should be shown initially in the scan path editor
by setting initialScaleUnit in TwoAxisSpiralModel.

Change-Id: I2a9b0eb808d4c07d32446bba4d7a3943a750a606