DAQ-2835 Enable PowerMock tests
A few things required, mostly adding PowerMockBase which makes a
modification to the classpath. Due to the particular classes which
are mocked by PowerMock in this plugin, and the consequences this causes
for the runtime state of class loading, it is required that all PowerMock
tests in this plugin extend this class.

Change-Id: Ib71e7699914708ab9c220c20b21a25b795b41281