K11-603 Simplify and generalise Experiment Structure jobs framework
'Point and Shoot' acquisitions will produce a NeXus file that links to
individual scan files in the same way as the experiment 'index' file
links to individual acquisition files.

ExperimentController therefore now supports a tree structure instead of
a fixed two-level experiment-acquisition. A sub-node is added by calling
startMultipartAcquisition, which has a corresponding close method.

IndexFileCreationRequest is also changed to a general NodeFileCreation-
Request which can be used to compose any node in the tree.
To achieve generality the external links in the created file are now
under a single NXentry, where before they were directly under the NXroot
element.

This change also fixes a subtle bug which is that the file created by
NexusFileBuilder.createFile() was never closed.

Change-Id: Icca787692b2480bd976b7e0a87d40588098d1c86