K11-612 Refactoring the Absorption Tool to use multiple ROI
The AbsorptionComposite is in charge to display two ROI. The scientist uses one ROI to identify a dark area another for a bright area.
This refactoring adds to the ContextMenu in the IPlottingSystem associated with a camera stream (BrightnessProcessorBuilder.ProcessorBuilder.createRegion()).

AbsorptionComposite is then listening for ROIChangeEvents to update the table where values are displayed.

Change-Id: I50998aa5edc9f7f7b6d39ed6613e04d7d1f73e2b