DAQ-2886 (9.16) Restore ScannableGroup functions removed as part of DAQ-2327
The getGroupMember function was removed from ScannableGroup, and the
getGroupMemberNames function was removed from both ScannableGroup
and IScannableGroup as part of DAQ-2327, breaking backwards compatibility.

The getGroupMemberNames function was moved to a new interface,
IScannableGroupNamed, renamed as getGroupMembersNamesAsArray
and implemented only in ScannableGroupNamed.

The getGroupMember function was moved to the new IScannableGroupNamed
interface, renamed as getGroupMemberByName and also only implemented in
ScannableGroupNamed.

The changes to the interface meant that ScannableGroupNamed was not a
drop in replacement for ScannableGroup.

This commit represents the minimum of changes required to safely restore
functionality required by beamline scripts during this run, and further tidy up
and deduplication will be required once this commit is merged to master.

Change-Id: Ieb5135a9ad191548e3b9e628b1a92ffac85361b4