I08-341 Use link instead of copy/paste in CalibrationFrameCollector
When an instance of this object is included in a scan a single-frame
scan is performed at the start. This change changes the behaviour so that
instead of copying and pasting the dataset into the main scan file an
external link is used instead.

This should avoid any unintended data type conversion that could occur
when the dataset is read in Java.

Change-Id: I94d4164950d1df99fc039b07c09eeebba938f743