DAQ-2893 Add GdaJythonBuiltin annotation and parser
This allows static functions to be added to the Jython namespace as
builtin functions so that they do not need to be imported in every
script that requires them.

The wrapper around the functions also generates ducumentation based on
the method signature (via reflection) and a custom doc string passed to
the annotation.

If a method is overloaded, the methods are grouped together and the
single builtin function delegates to the correct method based on the
arguments passed at runtime. The doc strings for the individual methods
are combined so that `help(builtin_method)` still provides useful information.

eg in a class Commands

    public class Commands {
        @GdaJythonBuiltin("Add together two integers")
        public static long add(int i, int j) {
            return i + j;
        }
        @GdaJythonBuiltin("Add multiple doubles")
        public static double add(double... values) {
            return DoubleStream.of(values).sum();
        }
    }

the builtin function `add` would combine both methods and the docstring
would list both options. With Jython's built in help function, this
would give:

    >>> help(add)
    Help on built-in function add:

    add(...)
        add(double...) -> double
            Add multiple doubles
        add(int, int) -> long
            Add together two integers

        from com.example.Commands

    >>>

Builtins can be created directly by passing a name and list of methods
to the constructor or a class can be parsed and all available methods
returned.

Change-Id: I01fc2d167bae993432d433d15b6d7e071e9a3dc7