Add utility method for adding Jython objects as observers
When a Jython objects is added as an observer of a Java object on the
server, a reference is held outside the Jython namespace and can live on
after the namespace has been reset. If an observer is added in
localStation, multiple instances can end up all being active at once.

With reset hooks it is possible to deregister observers when the
namespace is reset but from Jython this is awkward as the closure may no
longer have the correct object referenced. To safely add a Jython
observable, the following is needed

    >>> source.addIObserver(listener)
    >>> add_reset_hook(lambda src=source,
                obs=listener: src.deleteIObserver(obs))

This is awkward and prone to errors. With the new utility method this is
reduced to

    >>> add_jython_observer(listener, source)

Change-Id: I6722b211fe90de28abfccd5349ca755c526503ba