[DAQ-2899] remove ScanDataModel and ScanDeviceModel
These allow a custom mapping between the dimensions of a dataset
belonging to a provided nexus device and the dimensions of the primary
dataset of a detector when constructing an NXdata group.

In practise, this model is never used. Removing it makes it easier to
refactor the code to build a nexus file, in preparation for moving it
down to org.eclipse.dawnsci.nexus

Change-Id: Iaf903dc8ef6423070267607028ae6abc522d59e5