DAQ-2867 More consistent JavaDoc and treatment of filePath in ScanRequest
The filePath field in ScanRequest was documented as:
   Part or all of the file path to be used for this scan.
but, in ScanProcess.setFilePath(), any non-null value was treated
as the full path to the output file.

This changes the behaviour to treat only a path ending in "<a>.<b>" as a
full file path (as before) and anything else as a directory
specification (as per JavaDoc). It also makes this more explicit in
the JavaDoc.

Also moves all ScanRequest constructors to the start of the class
and deletes redundant call to super() constructor.

Change-Id: Ie8c64d7d25f56a1927f51a03e6add7cb4611c2a0