DAQ-1980 Add Logback filter for source process name
Abandoning the LogServer means that all client events are sent to the
server's logging context so that they are written to a combined file.
This means that other appenders that are designed to be local to the
server only (console appender and Graylog appender) should filter out
non server log messages.

Change-Id: I1ad674677f58d9b7aa96fa491a7e923b3201c323