DAQ-2944 Add mode specific handling to SpringMotorDefinitionParser
Motor parser now behaves very similarly to the previous version except
now attributes prefixed with mode- will only be applied when GDA is
running in that mode.

If an attribute relates to a top level field, it is set as a property on
the ScannableMotor. All other attributes are set on the controller bean
without being checked. This allows fields that are not on the Motor
interface to be set and for future implementations to require no
additions to the parser class.

Change-Id: Ief55b56e7f7e2342da529408d27758d158a0ffd4