[DAQ-2932] use metadata appenders for scannables and metadata scannables
but only for scannables that do not have an entry in the location map.
This is because the location map allows each field of a scannable to be
written to any location, so there may not be a parent group for the
device. This could lead to potential clashes with metadata for other
devices.

Change-Id: I70853e91152a0949d1ef44296a734d1a36a98a26