DAQ-2975 Revert "DAQ-2038 Remove usage of deprecated API"
This reverts commit 5120b6d9810bf1fe3d0ba730a9e07ad3425823ee.

Reason for revert: This change causes indeterminate stream failures
with specific MJPEG streams (in particular RGB1 streams from MX OAV
IOCs).

Specifically the failure is caused by ImageIO inferring an incorrect
reader type e.g. BMPImageReader which fails to decode. It's not clear
whether this is caused by a) an inconstant stream from EPICS, b)
oversights/bugs in the handling of the stream (before this method is
reached) or c) flaws in ImageIO.

A workaround is possible by enforcing use of a JPEGImageReader however
as there is no requirement to remove Oracle JDK 8 usage on gda-9.17
it seems sensible to revert this change here and take the opportunity
to remove the API, reactor and improve the class on master.

Change-Id: I8e384e59b2971ceb398d38578bc507c427a05df1