Reduce the default update interval for ThreadlessDummyMotor
Currently, the default update interval is 500ms. For long moves this is fine,
but for short moves, or quick motors, the minimum time was half a second.

A much higher update frequency allows higher resolution of move times and
given a thread is only active during moves, does not increase idle resource
use.

This has the effect of increasing the default speed of
ThreadlessDummyMotors but given the previous default was 0.2/s (in
whatever units it used), this is also a good thing in most cases.

Change-Id: I76dc21402ecf29b28300f86a04782db1a23f61a9