DAQ-2993 Add new class to convert Mjpeg stream to Jpeg bytes
MjpegInputStream takes an InputStream and allows byte arrays containing
Jpeg data to be read from it.

Mjpeg header lines are read to find Content-Length which allows reading
the bytes in one pass. If Content-Length is not present reading can
still occur by looking ahead to find the Jpeg end marker. This resolves
GDA-3129.

Change-Id: I2f93ea69748d894433c8814411bb3dd921a95d4c