[DAQ-2568] retry after InterruptedException when calling malcolm method
Methods that send a message to the actual malcolm device and wait for a
reply should not support cancellation through the interrupt method.
This could cause the next scan to start while malcolm is still busy.
Most often this will be the run method, as that is the only
long-running method. The correct way to cancel a run is to call abort
on the MalcolmDevice.

Change-Id: I744cb20fca6ba7407a467da320761cb5074bbc99