DAQ-2994 Workaround for Logback scheduled executor limitation
Logback uses a ScheduledThreadPoolExecutor that has a fixed pool size
of 8 threads. Since two threads are used to listen for clients (GDA
clients and log panels) this only leaves 6 threads for clients
themselves to use. If connection events are queued this leads to an
overflow of CLOSE_WAIT sockets.

The method added here allows a monitoring of the queue and manage
the thread count as required.

Change-Id: Ifae2479b4a0792fc1eb714ee6ec4007489093224