DAQ-2993 Move image reading call into decode task
Converting from bytes to BufferedImage takes time (ImageIO takes longer
than the deprecated api). It is also a form of decoding so move it into
the decode task that is submitted to a queue.

This allows the relatively fast task of getting byte arrays of the
stream to proceed as fast as it can to keep up with the stream. The
queue should be responsible for dropping frames if there is a backup.

Change-Id: Ic2f6f867b35115c309d9270afe0f74095820e003