DAQ-2370 Deprecate Sixd data writers
SixdNexusDataWriter is not referenced anywhere and is likely missing
many features/fixes found in NexusDataWriter.

SixdSrsDataWriter is almost a complete duplicate of SrsDataFile apart
from "CustomerisedUserCommand" which is a way to write a scan command
into the data file. This may need to be added to SrsDataFile or
find a different way to get the scan command.

SixdSrsDataWriter is possibly still used on I06.

Change-Id: I8bafb401c8c1a79bd67cb8223b72abb34055aeba