Make RemoteCommandRunner stoppable
Commands started via a RemoteCommandRunner could not be stopped. If the command
was run using the exec Jython methods, the command would also be unstoppable via
StopAll command.

NB In the case of SingleCommandRunner, if a command ignores thread
interruptions, it may be possible for the command to continue running after
stop has been called. In this case, any further exceptions will not be
logged and additional commands may be run.

Change-Id: I2017d1b7f6bfbb836f9a4d7e45e22f0e809ff9a3