DAQ-704 Remove scannable overwriting check
Checks in the translator were fragile and caused problems in valid commands.
It was also possible to workaround the checks creating more difficult to
maintain code. After further discussion, this protection is something that
we no longer require.

Any workarounds can now be replaced with straightforward assignments. Any
scripts which used the  contextmanager will continue to work
for now but should be reverted to standard assignments before 9.20 when
it will be removed.

Change-Id: Id712e9a6766866a8ccb9cc3bbc4680081af8beeb