K11-638 Diffraction stream should have axis calibrated on KB mirrors
-- BEAM TO CAMERA MAPPING --
- Simplified beam-camera mapping using a grid scan based on
CameraMappingBean and BeamMappingStateContext

- BeamCameraCalibrationComposite provides the GUI interface to 1)
start/abort the mapping 2) report the advance beam drivers

- the user starts the mapping from BeamCameraCalibrationComposite which
starts the following workflow --> BeamMappingStateContext (calibration
state start/stop) --> BeamCameraMapping (Spring bean associated with the
beam motors) --> BeamCameraMappingHelper(does the real mapping)

- When a new mapping is available CameraHelper.addBeamCameraMap 1) a new
BeamCameraMap is set in the relevant ICameraConfiguration 2) publishes a
BeamCameraMappingEvent.

- BeamCameraMappingEvent listener is BeamCameraCalibrationComposite
which uses DrawCameraMappingArea to draw a rectangle on the live stream

-- INJECT MAPPING INTO PLOTTING-SYSTEM--
- The LiveStreamBackgroundAction allows the user to start or stop the
imaging stream for the diffraction perspective. When a BeamCameraMap is
available, LiveStreamBackgroundAction instantiate an instance of
BeamMappingSupport which is able, through AxesAction, to draw or remove
additional axes, calibrated on the given map, on the existing  stream.

- BeamMappingSupport contains a listener which may be activate
externally (see LiveStreamBackgroundAction.updateActionGUI). This
listener catches events like "press b +
click mouse" on the plotting system and moves accordingly the beam using
the available BeamCameraMap

Change-Id: I7b69b935c82363ed8cc12508935bfa79e3046259