[DAQ-2844] make caching optional in ScannablePositioner
this is required to move to subsequent positions that have different names

Note: the if test in ScannablePositioner.setPosition was inverted,
and a simpler fix would have been to correct it. However, it seemed
to make more sense simply to not cache at all rather than cache and
then clear the cache each time.

Change-Id: I74a9c0f41f5ccecfb4d7a04b8e399b54a344cc62