DAQ-3069 Change threading in FocusScanResultPage
The map update was using the file controller's update to refresh the
data. The GUI was using the scan bean update to update the UI.

Then scan bean finish (which cleaned up the repeating executor) came in
before the last data update. So the last data update re-initialised
everything, including clearing the plot.

Since the map updates from the controller come every 2s or slower it
seemed pointless to keep an executor running periodically just to
refresh the file. So we removed this which also removed the need for
the executor clean up, fixing the issue

Change-Id: I617800edc28f336011acc894184a7680aa6e4948