DAQ-3079 Reduce logging level relating to restricted tp wildcard imports
The lack of wildcard import restriction is unlikely to cause issues
for Jython usage so reduce the logging level. The reasoning behind this
code will be investigated before removal later on.

It was reasonable to be logged at error previously as lack of
target_platform_artifacts.xml could indicate a failed server
build. However this file is not produced by the Maven build.

Change-Id: I0113a71cdc55306d83dbfefede8a748aab9fdfdd