Convert Spring motor parser to use new utilities
This brings a minor change for motors that do not use default classes
(eg epics motor in dummy mode or dummy motor in live mode). Where
previously the attribute uses was modeClass, it is now mode-class. This
allows a single approach to be used to parse all attributes and
duplicated code to be reduced.

Change-Id: I1d9f62a7f0865a094f2a2bbf49e87f53114668fb