[DAQ-3075] split ScannableNexusWrapper to create ScannableNexusDevice
ScannableNexusDevice wraps a gda.device.Scannable to implement both
IScannable, the solstice scanning interface for scannables, and
INexusDevice, to provide a nexus object (GroupNode) to the new
NexusScanFile based nexus writing framework.

In order to run a old style scan with the new nexus writing framework,
we only want the nexus writing part of that. Splitting the nexus writing
parts of ScannableNexusWrapper into its own class, ScannableNexusDevice
not only allows us to do that, but also is a better design anyway,
according to the single-responsibility principle, as each class now
only does one thing.

Change-Id: I7e7a5b183cfcacdafaf5c7b29f36430d156d1381