[DAQ-3158] remove SampleData, as it is not used
The ScanRequestConverter until recently did not set the SampleData.
When it was changed to do this, it was discovered that ScanProcess
adds this to the nexus file name, which we do not want. Removing this
use of SampleData makes it unused, meaning that it can be removed.

Note that the sample name and description are added to the ScanRequest
as part of a ScanMetadata object. This is an object that contains a
key-value map and a type, one of SAMPLE (this case), ENTRY or INSTRUMENT.
This is a more general way or adding metadata to scans, and is used to
add metadata to the corresponding groups of the nexus file. If
SampleData were to be used for this purpose also, there would be the
the potential for a clash and metadata being lost.

Change-Id: I8bb1c56c1a9f6a5c4aa2ad30537fbe34caa4e8ee