K11-894 Inject positions in DarkCalibration and FlatCalibration
- Defines a FindableService available in Spring for @Autowire
- Defines a DevicePositionDocumentService to generate positions document
used to either request an action or to report a status. The service core
is implemented using a Chain of responsibility pattern so any new device
can be easily added implementing DeviceHandlerBase
- The StageController already was responsible to store different
positions (start, end, outOfBeam). To limit the refactoring of the
stageController to a successive ticket, it adds a parallel structure to
support the persistence of DevicePositionDocument for different
positions
- The ScanningAcquisitionController, already aware of the
StageController, can update DarkCalibration and FlatCalibration, in the
ScanningAcquisition, before save or run the acquisition.
- Unit tests are added to support the change
- Still dark/flat position are missing

Change-Id: I98173599659afac5be75007dc259cd78ccb1a37c