I08-355 Allow specification of Malcolm device for dark field collection
This allows a different Malcolm device to be used for the dark field
collection than the one used for the "main" scan.

In I08-1, this allows us to use standard Malcolm devices (including
fast shutter control) for most scans, and a special Malcolm device
(with no shutter control) only for dark frame collection.

Note that this requires some rather hacky code to set the exposure time,
since the ScanModel at the time it is passed to this class has
inconsistent exposure times in the various models.
See comments in the code, and also DAQ-3165

Change-Id: I94d93bc92d3e0c533da0944827b710d495ce9459