K11-872 Enforces ScannableTrackDocument axis
- Until now DetailsLabelProvider classes have used
ScannableTrackDocument by their index. The new axis property gives a
reference that makes the index not necessary. Beware that this change
will break the existing acquisition configurations as those do not
contain the axis property.
- ScanningParametersHelperBase implements new methods based on
ScannableTrackDocument axis property and deprecate those based on
indexes
- Prepare for the remove of
ScanningParametersHelperBase.getScannableTrackDocumentBuilder(int)
reducing the methods depending on it

Change-Id: I9045fed0aae9a7dc01ed56d9a128e223804cdf5f