simplify call to writeCounterTimer and remove 2 arg method
Previously the method writeDetector, in the case of a counterTimer
(where detector is not a NexusDetector and the det.getExtraNames()
returns an array of (size > 1), called extractDoubleData to get the data
to add. If this was not null, it called writeCounterTimer with the
detector name and data. If not it called another version that only takes
the detector name was called. But all this did is call the same
extractDoubleData method that previously returned null, which must just
return null again.

This change removes the previous one arg version of writeCounterTimer,
removes the data argument from previous two arg version, adding a call
to get the data at the top of that method.

Change-Id: I5925b570d0d1f2af94dd217cd7a976a19fc46d73