use Metadata object to get beamline name in NeXusUtils
NexusDataWriter gets the beamline from the metadata object, using
metadata.getMetadataValue(), with LocalProperties.GDA_INSTRUMENT as the
fallback property, and "base" as the fallback value. NeXusUtils did
not, which meant that these properties could diverge, especially in tests.
In each class the beamline is used to create a NumTracker. If this
happened, the entry identifier added to a scan would always be 0.

This change fixes NeXusUtils to get the beamline in exactly the same
way that NexusDataWriter does.

Change-Id: I8138bbea8ff2312f60b232de50683e378c56f695