DAQ-2650 Monitor the health of the beamline
This allows configuration of a number of scannables that are deemed
relevant to the overall health of the beamline. This allows you to
specify, for example, that for the beamline to be "healthy", there must
be a certain current in the ring, and the front-end shutter must be
open.

If a condition is configured as "critical", it will be regarded as an
error if it is not met, otherwise it will be regarded as a warning.

Change-Id: I0134e46c9b94e322fc8b201d2eebc7218e6c7eab