DAQ-3180 Refactor ScannableUtils objectToArray
Reduce number of extra arrays that need to be allocated by
ArrayUtils.add calls. Create objectToDouble method to reduce duplicated
conversion code for individual values.

Enable the ignored test for the case that now works (null -> empty).
Arrays of objects that have a toString that returns a parsable double
can now be converted so modify the test for that case. Previously mapped
to array of nulls whereas a single value of the same class would map to
an array of length 1 with the parsed value.

Change-Id: I1dd544c3d8cc5e55cb3b77db8a8c65dc4408e886