[DAQ-3179] split out makeNexusDetector method from  makeGenericDetector
Looking at the code for this method in more detail, most of it is
different for the two cases:
 - detector instanceof NexusDetector and detectorData is non-null
 - detector is not NexusDetector and detectorData is null

Therefore it makes sense to split this method in two. The code that
is duplicated is already duplicated multiple times for the other cases,
e.g. makeCounterTimer, makeFileCreatorDetector, etc. So further
refactoring could be done, but would be too time-consuming at present.

Change-Id: Ib514f087f51e7d958ab3bb10c9a92096e0161937