I19-456 Adds option to reorder or suppress the StopAll button placement
Previous DAQ-2637 replaced a configurable StopAll button with a
hard-coded version for ease of programmatic placement. Unfortunately,
this removed the option of suppressing the placement of the default
StopAll button in the status line. (eg i19-2)

Properties added to hide or position left, both default false in absence
of the opt-in properties.

Change-Id: Ic06d16cf42dcfb7463553f47d2cd5991e3dc282f