DAQ-3364 : Add check to confirm change in Visit during client startup
When starting GDA client as staff, a dialog box can now be displayed if
the currently set visit is different from the one the client is about to
set for the new user. This is to help avoid (accidentally) replacing a
user set visit directory with commissioning directory when running
multiple clients.
This behaviour can be switched on by setting gda.client.confirm.visit to
true in plugin_customization.ini file.

Change-Id: I76a5dea9b5982771aa724922e6f89ae5669fcf77