Remove unreliable aspects of ScannableUtilsTest
Several parts of this test relied on the global state of Jython
internals. As such it was liable to fail for unrelated reasons occasionally.
The removed aspects do appear to work as expected in practice and are very
rarely used features that do not justify further investigation.

Py.newDecimal imports the decimal module and instantiates a new instance
from it. 'Normal' float and int values are not affected.

Py.Ellipsis is a singleton object that caused problems when trying
to convert to a double. This was originally added as an instance of an
arbitrary non-double python object and is not a case that is expect to
ever be used.

Change-Id: I1ff92ded7983b4052d3686d646330979eea5e04d